refactor(engine): move TypedEventImpl to streamprocessors package
