+++ engine/src/test/java/io/camunda/zeebe/engine/processing/streamprocessor/TypedEventSerializationTest.java
@@ -19,6 +19,7 @@ import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.DeploymentIntent;
+import io.camunda.zeebe.streamprocessor.TypedEventImpl;
 import io.camunda.zeebe.test.util.JsonUtil;
 import io.camunda.zeebe.util.collection.Tuple;
 import org.agrona.DirectBuffer;
